Adding a Helm chart for MuShop
Each component of the project is a subchart with it's own settings and it can be enabled/disable from the main chart by setting the `enabled` flag to appropriate value. By default, all components are enabled and a dev version of the chart is installed (i.e. ingress without the tls certs/hosts).

In order for the chart to install, you will need to provide the DB (passwords) and the OCI information - previously, you would manually have to create a secret with this information. Here's a sample command for installing the chart:
```
helm install mushop --name mymushop \
    --set catalogue.secrets.oadbPassword=xxxxxx \
    --set carts.secrets.oadbPassword=xxxxxx \
    --set orders.secrets.oadbPassword=xxxxxx \
    --set secrets.oci.compartmentId=<your compartment id> \
    --set secrets.oci.tenantId=<your tenant id> \
    --set secrets.oci.userId=<your user id> \
    --set secrets.oci.region=<your region> \
    --set secrets.oci.apiKey="$OCI_KEY"
```

Note: in order for the installation work in Prod, you need to create the following secrets first:

letsencrypt-prod
mushop-tls-prod-secret
This can be added later once we have the CI/CD for prod set up.

Resolves #71