removed package naming localtion
As requested, removed the github.com/repo reference on the package name as the solution does not have a definitive house.

This is not recomended but does not affects the usage.

Affects the dependecies extraction for 3rd-party license, as this is done from the binary and not from the go.mod. So, will need to change every time that do license extraction or change the extractor.

As soon get the definitive repo, should change to the recommended way to name packages.

P.S.: all other services in go are using the github.com/repo naming.