Adding the newsletter-subscription function
I am also removing the mock mode for the function. If the URL is not provided/defined, we will just return a 200. That way we don't need to explicitly specify the mock mode for the function. Later if we want to or if we think it makes sense, we could modify the way mock mode is detected to also check if the env variables are not provided, to default to mock mode or something similar.